samples/azure_functions/requirements.txt
azure-functions<2
slack_bolt